requirements.txt
fastapi
uvicorn
torch
pillow
git+https://github.com/openai/CLIP.git
opencv-python